softmax/pyproject.toml
[project]
name = "softmax"
version = "0.1.0"
description = "investigations for various softmax kernels"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "ninja>=1.13.0",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pyyaml>=6.0",
    "torch>=2.10.0",
    "triton>=3.6.0",
]

# Only build the C++ extension; do not treat csrc/images/other dirs as Python packages
[tool.setuptools.packages.find]
include = []

[build-system]
# ninja: PyTorch BuildExtension uses it by default; without it on PATH it falls back to slow distutils.
requires = ["setuptools", "wheel", "ninja>=1.13.0", "numpy>=2.4.3", "torch>=2.10.0", "pyyaml"]
build-backend = "setuptools.build_meta"
